requirements.txt
requests
logger
azure-communication-email
psycopg2-binary == 2.9.9